Fix modal focus management & Cleanup CustomNodeModal
