+++ common/components/Header/components/CustomNodeModal.scss
@@ -1,0 +1,15 @@
+.CustomNodeModal {
+  .flex-wrapper {
+    margin: 0px -8px;
+    > .input-group {
+      margin: 0px 8px;
+      > .input-group-input {
+        width: 100%;
+      }
+    }
+  }
+  input[type='checkbox'] {
+    margin-right: 1rem;
+    margin-bottom: 1rem;
+  }
+}
