+++ common/components/Header/components/CustomNodeModal.tsx
@@ -13,17 +13,10 @@ import {
 } from 'selectors/config';
 import { CustomNode } from 'libs/nodes';
 import { Input } from 'components/ui';
+import Dropdown from 'components/ui/Dropdown';
+import './CustomNodeModal.scss';
 
-const CUSTOM = 'custom';
-
-interface InputProps {
-  name: string;
-  placeholder?: string;
-  type?: string;
-  autoComplete?: 'off';
-  onFocus?(): void;
-  onBlur?(): void;
-}
+const CUSTOM = { label: 'Custom', value: 'custom' };
 
 interface OwnProps {
   isOpen: boolean;
@@ -56,7 +49,7 @@ interface State {
 type Props = OwnProps & StateProps & DispatchProps;
 
 class CustomNodeModal extends React.Component<Props, State> {
-  public state: State = {
+  public INITIAL_STATE = {
     name: '',
     url: '',
     network: Object.keys(this.props.staticNetworks)[0],
@@ -67,6 +60,14 @@ class CustomNodeModal extends React.Component<Props, State> {
     username: '',
     password: ''
   };
+  public state: State = this.INITIAL_STATE;
+
+  public componentDidUpdate(prevProps) {
+    // Reset state when modal opens
+    if (!prevProps.isOpen && prevProps.isOpen !== this.props.isOpen) {
+      this.setState(this.INITIAL_STATE);
+    }
+  }
 
   public render() {
     const { customNetworks, handleClose, staticNetworks, isOpen } = this.props;
@@ -89,7 +90,13 @@ class CustomNodeModal extends React.Component<Props, State> {
     ];
 
     const conflictedNode = this.getConflictedNode();
-
+    const staticNetwrks = Object.keys(staticNetworks).map(net => {
+      return { label: net, value: net };
+    });
+    const customNetwrks = Object.entries(customNetworks).map(([id, net]) => {
+      return { label: net.name + ' (Custom)', value: id };
+    });
+    const options = [...staticNetwrks, ...customNetwrks, CUSTOM];
     return (
       <Modal
         title={translate('NODE_Title')}
@@ -98,146 +105,132 @@ class CustomNodeModal extends React.Component<Props, State> {
         handleClose={handleClose}
         maxWidth={580}
       >
-        <div>
-          {isHttps && <div className="alert alert-warning small">{translate('NODE_Warning')}</div>}
+        {isHttps && <div className="alert alert-warning small">{translate('NODE_Warning')}</div>}
 
-          {conflictedNode && (
-            <div className="alert alert-warning small">
-              You already have a node called '{conflictedNode.name}' that matches this one, saving
-              this will overwrite it
+        {conflictedNode && (
+          <div className="alert alert-warning small">
+            You already have a node called '{conflictedNode.name}' that matches this one, saving
+            this will overwrite it
+          </div>
+        )}
+
+        <form className="CustomNodeModal">
+          <div className="flex-wrapper">
+            <label className="col-sm-9 input-group flex-grow-1">
+              <div className="input-group-header">Node Name</div>
+              <Input
+                className={`input-group-input ${this.state.name && invalids.name ? 'invalid' : ''}`}
+                type="text"
+                placeholder="My Node"
+                value={this.state.name}
+                onChange={e => this.setState({ name: e.currentTarget.value })}
+              />
+            </label>
+            <label className="col-sm-3 input-group">
+              <div className="input-group-header">Network</div>
+              <Dropdown
+                className="input-group-dropdown"
+                value={network}
+                options={options}
+                clearable={false}
+                onChange={e => this.setState({ network: e.value })}
+              />
+            </label>
+          </div>
+
+          {network === CUSTOM.value && (
+            <div className="flex-wrapper">
+              <label className="col-sm-6 input-group input-group-inline-dropdown">
+                <div className="input-group-header">Network Name</div>
+                <Input
+                  className={`input-group-input ${
+                    this.state.customNetworkId && invalids.customNetworkId ? 'invalid' : ''
+                  }`}
+                  type="text"
+                  placeholder="My Custom Network"
+                  value={this.state.customNetworkId}
+                  onChange={e => this.setState({ customNetworkId: e.currentTarget.value })}
+                />
+              </label>
+              <label className="col-sm-3 input-group input-group-inline-dropdown">
+                <div className="input-group-header">Currency</div>
+                <Input
+                  className={`input-group-input ${
+                    this.state.customNetworkUnit && invalids.customNetworkUnit ? 'invalid' : ''
+                  }`}
+                  type="text"
+                  placeholder="ETH"
+                  value={this.state.customNetworkUnit}
+                  onChange={e => this.setState({ customNetworkUnit: e.currentTarget.value })}
+                />
+              </label>
+              <label className="col-sm-3 input-group input-group-inline-dropdown">
+                <div className="input-group-header">Chain ID</div>
+                <Input
+                  className={`input-group-input ${
+                    this.state.customNetworkChainId && invalids.customNetworkChainId
+                      ? 'invalid'
+                      : ''
+                  }`}
+                  type="text"
+                  placeholder="1"
+                  value={this.state.customNetworkChainId}
+                  onChange={e => this.setState({ customNetworkChainId: e.currentTarget.value })}
+                />
+              </label>
             </div>
           )}
 
-          <form>
-            <div className="row">
-              <div className="col-sm-7">
-                <label>{translate('NODE_Name')}</label>
-                {this.renderInput(
-                  {
-                    name: 'name',
-                    placeholder: 'My Node'
-                  },
-                  invalids
-                )}
-              </div>
-              <div className="col-sm-5">
-                <label>Network</label>
-                <select
-                  className="form-control"
-                  name="network"
-                  value={network}
-                  onChange={this.handleChange}
-                >
-                  {Object.keys(staticNetworks).map(net => (
-                    <option key={net} value={net}>
-                      {net}
-                    </option>
-                  ))}
-                  {Object.entries(customNetworks).map(([id, net]) => (
-                    <option key={id} value={id}>
-                      {net.name} (Custom)
-                    </option>
-                  ))}
-                  <option value={CUSTOM}>Custom...</option>
-                </select>
-              </div>
-            </div>
+          <label className="input-group input-group-inline-dropdown">
+            <div className="input-group-header">URL</div>
+            <Input
+              className={`input-group-input ${this.state.url && invalids.url ? 'invalid' : ''}`}
+              type="text"
+              placeholder="https://127.0.0.1:8545/"
+              value={this.state.url}
+              onChange={e => this.setState({ url: e.currentTarget.value })}
+              autoComplete="off"
+            />
+          </label>
 
-            {network === CUSTOM && (
-              <div className="row">
-                <div className="col-sm-6">
-                  <label className="is-required">Network Name</label>
-                  {this.renderInput(
-                    {
-                      name: 'customNetworkId',
-                      placeholder: 'My Custom Network'
-                    },
-                    invalids
-                  )}
-                </div>
-                <div className="col-sm-3">
-                  <label className="is-required">Currency</label>
-                  {this.renderInput(
-                    {
-                      name: 'customNetworkUnit',
-                      placeholder: 'ETH'
-                    },
-                    invalids
-                  )}
-                </div>
-                <div className="col-sm-3">
-                  <label>Chain ID</label>
-                  {this.renderInput(
-                    {
-                      name: 'customNetworkChainId',
-                      placeholder: 'e.g. 1'
-                    },
-                    invalids
-                  )}
-                </div>
-              </div>
-            )}
+          <label>
+            <input
+              type="checkbox"
+              name="hasAuth"
+              checked={this.state.hasAuth}
+              onChange={() => this.setState({ hasAuth: !this.state.hasAuth })}
+            />
+            <span>HTTP Basic Authentication</span>
+          </label>
 
-            <div className="row">
-              <div className="col-sm-12">
-                <label>URL</label>
-                {this.renderInput(
-                  {
-                    name: 'url',
-                    placeholder: 'e.g. https://127.0.0.1:8545/',
-                    autoComplete: 'off'
-                  },
-                  invalids
-                )}
-              </div>
+          {this.state.hasAuth && (
+            <div className="flex-wrapper ">
+              <label className="col-sm-6 input-group input-group-inline-dropdown">
+                <div className="input-group-header">Username</div>
+                <Input
+                  className={`input-group-input ${
+                    this.state.username && invalids.username ? 'invalid' : ''
+                  }`}
+                  type="text"
+                  value={this.state.username}
+                  onChange={e => this.setState({ username: e.currentTarget.value })}
+                />
+              </label>
+              <label className="col-sm-6 input-group input-group-inline-dropdown">
+                <div className="input-group-header">Password</div>
+                <Input
+                  className={`input-group-input ${
+                    this.state.password && invalids.password ? 'invalid' : ''
+                  }`}
+                  type="password"
+                  value={this.state.password}
+                  onChange={e => this.setState({ password: e.currentTarget.value })}
+                />
+              </label>
             </div>
-
-            <div className="row">
-              <div className="col-sm-12">
-                <label>
-                  <input
-                    type="checkbox"
-                    name="hasAuth"
-                    checked={this.state.hasAuth}
-                    onChange={this.handleCheckbox}
-                  />{' '}
-                  <span>HTTP Basic Authentication</span>
-                </label>
-              </div>
-            </div>
-            {this.state.hasAuth && (
-              <div className="row">
-                <div className="col-sm-6">
-                  <label className="is-required">Username</label>
-                  {this.renderInput({ name: 'username' }, invalids)}
-                </div>
-                <div className="col-sm-6">
-                  <label className="is-required">Password</label>
-                  {this.renderInput(
-                    {
-                      name: 'password',
-                      type: 'password'
-                    },
-                    invalids
-                  )}
-                </div>
-              </div>
-            )}
-          </form>
-        </div>
+          )}
+        </form>
       </Modal>
-    );
-  }
-
-  private renderInput(input: InputProps, invalids: { [key: string]: boolean }) {
-    return (
-      <Input
-        className={`${this.state[input.name] && invalids[input.name] ? 'invalid' : ''}`}
-        value={this.state[input.name]}
-        onChange={this.handleChange}
-        autoComplete="off"
-        {...input}
-      />
     );
   }
 
@@ -279,7 +272,7 @@ class CustomNodeModal extends React.Component<Props, State> {
     }
 
     // If they have a custom network, make sure info is provided
-    if (network === CUSTOM) {
+    if (network === CUSTOM.value) {
       if (!customNetworkId) {
         invalids.customNetworkId = true;
       }
@@ -316,7 +309,7 @@ class CustomNodeModal extends React.Component<Props, State> {
     const { network, url, name, username, password } = this.state;
 
     const networkId =
-      network === CUSTOM
+      network === CUSTOM.value
         ? this.makeCustomNetworkId(this.makeCustomNetworkConfigFromState())
         : network;
 
@@ -349,20 +342,10 @@ class CustomNodeModal extends React.Component<Props, State> {
     return customNodes[config.id];
   }
 
-  private handleChange = (ev: React.FormEvent<HTMLInputElement | HTMLSelectElement>) => {
-    const { name, value } = ev.currentTarget;
-    this.setState({ [name as any]: value });
-  };
-
-  private handleCheckbox = (ev: React.FormEvent<HTMLInputElement>) => {
-    const { name } = ev.currentTarget;
-    this.setState({ [name as any]: !this.state[name as keyof State] });
-  };
-
   private saveAndAdd = () => {
     const node = this.makeCustomNodeConfigFromState();
 
-    if (this.state.network === CUSTOM) {
+    if (this.state.network === CUSTOM.value) {
       const network = this.makeCustomNetworkConfigFromState();
 
       this.props.addCustomNetwork({ config: network, id: node.network });
