+++ common/components/ui/Input.scss
@@ -29,6 +29,9 @@
       color: rgba(0, 0, 0, 0.54);
     }
   }
+  &-dropdown {
+    margin-bottom: 1rem;
+  }
   &-input {
     width: 100%;
     border: 1px solid #e5ecf3;
