+++ common/components/ui/Modal/index.tsx
@@ -1,0 +1,74 @@
+import React, { PureComponent } from 'react';
+import { CSSTransition, TransitionGroup } from 'react-transition-group';
+import './index.scss';
+import ModalBody from './ModalBody';
+
+export interface IButton {
+  text: string | React.ReactElement<string>;
+  type?: 'default' | 'primary' | 'success' | 'info' | 'warning' | 'danger' | 'link';
+  disabled?: boolean;
+  onClick?(): void;
+}
+interface Props {
+  isOpen?: boolean;
+  title?: string | React.ReactElement<any>;
+  disableButtons?: boolean;
+  children: any;
+  buttons?: IButton[];
+  maxWidth?: number;
+  handleClose(): void;
+}
+interface ModalStyle {
+  width?: string;
+  maxWidth?: string;
+}
+
+const Fade = ({ children, ...props }) => (
+  <CSSTransition {...props} timeout={300} classNames="animate-modal">
+    {children}
+  </CSSTransition>
+);
+
+export default class Modal extends PureComponent<Props, {}> {
+  public componentDidMount() {
+    this.toggleScroll();
+  }
+
+  public componentDidUpdate() {
+    this.toggleScroll();
+  }
+
+  public toggleScroll() {
+    document.body.classList.toggle('no-scroll', !!this.props.isOpen);
+  }
+
+  public componentWillUnmount() {
+    document.body.classList.remove('no-scroll');
+  }
+
+  public render() {
+    const { isOpen, title, children, buttons, handleClose, maxWidth } = this.props;
+    const hasButtons = buttons && buttons.length;
+    const modalStyle: ModalStyle = {};
+
+    if (maxWidth) {
+      modalStyle.width = '100%';
+      modalStyle.maxWidth = `${maxWidth}px`;
+    }
+
+    const modalBodyProps = { title, children, modalStyle, hasButtons, buttons, handleClose };
+
+    return (
+      <TransitionGroup>
+        {isOpen && (
+          <Fade>
+            <div>
+              <div className="Modal-overlay" onClick={handleClose} />
+              <ModalBody {...modalBodyProps} />
+            </div>
+          </Fade>
+        )}
+      </TransitionGroup>
+    );
+  }
+}
