+++ common/sass/styles/flexbox.scss
@@ -7,6 +7,19 @@
     flex-wrap: nowrap;
   }
 }
+
+.flex-grow {
+  &-1 {
+    flex-grow: 1;
+  }
+  &-2 {
+    flex-grow: 2;
+  }
+  &-3 {
+    flex-grow: 3;
+  }
+}
+
 .flex-spacer {
   flex-grow: 2;
 }
